add proper HierarchyDbTests
